+++ board-frontend/html/scripts/Document.js
@@ -12,6 +12,7 @@ import Shape from "./canvas/canvasObjects/Shape.js";
 import {CanvasObjectType} from "./canvas/CanvasObjectType.js";
 import CanvasObjectWrapper from "./canvas/CanvasObjectWrapper.js";
 import SocketEventType from "./protocol/SocketEventType.js";
+import ClientMessageGetInvite from "./protocol/client/messages/ClientMessageGetInvite.js";
 
 const documents = new Map();
 let activeDocument = null;
@@ -117,7 +118,10 @@ socket.addMessageListener(ServerMessageType.UPDATE_CANVAS, (serverMessageUpdateC
     }
 });
 socket.addMessageListener(ServerMessageType.UPDATE_DOCUMENT, (serverMessageUpdateDocument) => {
-    documents.set(serverMessageUpdateDocument.id, new Document(serverMessageUpdateDocument.name, serverMessageUpdateDocument.id));
+    console.log(serverMessageUpdateDocument.shared);
+    documents.set(serverMessageUpdateDocument.id,
+        new Document(serverMessageUpdateDocument.name + (serverMessageUpdateDocument.shared ? "(shared)" : ""),
+            serverMessageUpdateDocument.id));
 });
 socket.addMessageListener(ServerMessageType.HANDSHAKE, (serverMessageHandshake) => {
     window.localStorage.setItem('id', serverMessageHandshake.id.toString());
@@ -139,22 +143,24 @@ socket.addEventListener(SocketEventType.OPEN, () => {
     } else {
         token = BigInt(0);
     }
-    socket.send(new ClientMessageHandshake(id, token));
-    let invite = document.location.href.substring(document.location.href.lastIndexOf("/") + 1);
-    if (invite !== '') {
-        try {
-            let bigint = BigInt(invite);
-            //socket.sendOpen(bigint); todo
-        } catch(e) {
-            console.error('improper invite', invite);
-        }
+    let invite;
+    let index = document.location.href.lastIndexOf('invite=');
+    if (index === -1) {
+        invite = '';
+    } else {
+        invite = document.location.href.substring(index + 7, index + 7 + 6);
     }
+    socket.send(new ClientMessageHandshake(id, token, invite));
 });
 
 const inviteButton = document.getElementById("inviteButton");
 
 inviteButton.addEventListener('click', (event) => {
-    console.log('invite');
+    socket.send(new ClientMessageGetInvite());
+})
+
+socket.addMessageListener(ServerMessageType.GET_INVITE, (serverMessageGetInvite) => {
+    window.alert('Join at localhost/index.html?invite=' + serverMessageGetInvite.code);
 })
 
 const MAX_TIME = 2000;