+++ board-frontend/html/scripts/protocol/client/messages/ClientMessageGetInvite.js
@@ -1,0 +1,16 @@
+import ClientMessageType from "../ClientMessageType.js";
+import ClientMessage from "../ClientMessage.js";
+
+export default class ClientMessageGetInvite extends ClientMessage {
+    constructor() {
+        super(ClientMessageType.GET_INVITE);
+    }
+
+    getBufferSize() {
+        return super.getBufferSize();
+    }
+
+    serialize(writer) {
+        super.serialize(writer);
+    }
+}