+++ board-frontend/html/scripts/protocol/client/messages/ClientMessageHandshake.js
@@ -2,10 +2,11 @@ import ClientMessageType from "../ClientMessageType.js";
 import ClientMessage from "../ClientMessage.js";
 
 export default class ClientMessageHandshake extends ClientMessage {
-    constructor(id, token) {
+    constructor(id, token, invite) {
         super(ClientMessageType.HANDSHAKE);
         this.id = id;
         this.token = token;
+        this.invite = invite;
     }
 
     getBufferSize() {
@@ -16,5 +17,7 @@ export default class ClientMessageHandshake extends ClientMessage {
         super.serialize(writer);
         writer.writeBigInt64(this.id);
         writer.writeBigInt64(this.token);
+        console.log(this.invite);
+        writer.writeString(this.invite);
     }
 }